refactor(NameValidator): 클래스 병합
